@@ -1,7 +1,7 @@
 {
   "name": "mini-html-extract-plugin",
   "author": "luics.xu@gmail.com",
-  "version": "0.1.3",
+  "version": "0.2.0",
   "license": "Apache-2.0",
   "description": "Generate HTML wrapper file(s) for specific entries.",
   "keywords": [
@@ -34,25 +34,26 @@
     "test:mocha": "mocha -r ts-node/register test/**/*.spec.ts"
   },
   "devDependencies": {
-    "@types/assert": "^1.5.1",
-    "@types/loader-utils": "^2.0.1",
-    "@types/mocha": "^7.0.2",
-    "@types/node": "^14.0.13",
-    "@types/react": "^16.9.36",
-    "@types/react-dom": "^16.9.8",
-    "@types/webpack": "^4.41.18",
-    "clean-webpack-plugin": "^3.0.0",
-    "mocha": "^8.0.1",
-    "nodemon": "^2.0.4",
+    "@types/assert": "^1.5.4",
+    "@types/loader-utils": "^2.0.2",
+    "@types/mocha": "^8.2.2",
+    "@types/node": "^15.9.0",
+    "@types/react": "^17.0.9",
+    "@types/react-dom": "^17.0.6",
+    "@types/webpack": "^5.28.0",
+    "clean-webpack-plugin": "^4.0.0-alpha.0",
+    "loader-utils": "^2.0.0",
+    "mocha": "^8.4.0",
+    "nodemon": "^2.0.7",
     "npm-run-all": "^4.1.5",
-    "raw-loader": "^4.0.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "ts-loader": "^7.0.5",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.5",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.11.0"
+    "raw-loader": "^4.0.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "ts-loader": "^8.0.0",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.2",
+    "webpack": "^5.38.1",
+    "webpack-cli": "^4.7.0",
+    "webpack-dev-server": "^3.11.2"
   }
 }
