@@ -57,7 +57,7 @@ module.exports = {
   resolve: { extensions: ['.ts', '.tsx', '.js'], },
   module: {
     rules: [
-      { test: /\.tsx?$/i, loader: 'ts-loader' },
+      { test: /\.tsx?$/i, loader: 'ts-loader', exclude: /node_modules/ },
     ]
   },
   devtool: false,
